Add a realloc failure error code
This error code signals that the allocator passed to the function could
not perform its task when used in a `realloc` fashion.

When this error code is returned, the `data` slot in the result struct
will get the value of the old allocation's pointer, so the callee can
clean up appropriately.