Fix returning the number of leftover bytes
The `void*` member of the result is used as an opaque type that could
carry not only a pointer value, but in this case the number of leftover
bytes as well.

Dereferencing it at the end of the parser meant that the null pointer
was passed to `memcpy`, instead of the address of the opaque pointer in
the callee's stack.