Remove the helpfully readable binary literals
Apparently this is not in the C language, because it "was rejected due
to lack of precedent and insufficient utility".

See: http://www.open-std.org/jtc1/sc22/wg14/www/C99RationaleV5.10.pdf

This is extremely dumb for a language used a lot for bit twiddling and
other similarly low level tasks.