Specify the `flags` argument to `mmap`
A value of 0 is not valid. All the possible flags are not really related
to reading, but `MAP_SHARED` seems to be the best option here. The file
is opened for read only, so no actualy writes will happen to this memory
region that has to be written back to the backing file.