Add fuzz testing for the parser
The fuzz testing uses corpus data that was generated (or seeded) from
the test assets.

This corpus data is stored in a persistent storage, which is
https://storj.io/ in this case. It provides free storage that is enough
for this purpose and it's a developer oriented service, easily
accessible using rclone.

The fuzzer runs for 2 minutes, which should be enough to find bugs if
there are any.

Note that the fuzzer requires the library under test (LUT) to be a
static library.

The compilation of the fuzzer is done using a superbuild, which can be
found in the fuzz directory. This superbuild uses ExternalProject to
connect the library project with the fuzzer's project.