Provide the tail block label with a statement
Labels may only be followed by statements, so this code had to be split
up in an earlier commit.

To fix an unreachable code warning, the original code was more or less
restored and wrapped in a do-while-zero loop to create a statement for
the label.

This doesn't quite achieve total warning free library code, but it gets
things closer to that.