Add an iterator checker macro for sanity checks
The ITER_CHECK macro asserts that the begin iterator is less than the
end iterator. This invariant comes from the API functions taking a
pointer + size_t size, from which begin and end are derived from. The
only faulty case that could occur is when a buffer size of 0 is passed
to the API functions.