Create API function to stringify `gif_result_code`
To achieve this stringification the gif_result_code enum was moved to a
header file to be on its own and a CMake function was written to read
this header file and output a source file.

This can be a little brittle, but Doxygen does not support X macros,
which would've been the most convenient way to achieve this. Moving the
enumerations out of the enum body makes Doxygen lose track of them and
output wrong documentation.