Add directory tree for IDE generators
There are IDEs out there that refuse to use the file system as a base to
display relevant sources for targets, so this commit implements source
groups for the main library target, which has source files nested in
directories.

The target_sources_grouped command is introduced to avoid having to use
variables in the root lists file to pass to both the library and the
source_group command.