Use `export(EXPORT)` to create a CMake package
export(EXPORT) is similar to install(EXPORT), but instead of making the
project available for reuse only after install, this command creates a
CMake package in the binary directory. This could be more convenient,
because now the artifacts exist only in one place, but on platforms
where RPATH is a thing, you will not have to specify it to make files
usable, because in the binary directory CMake makes sure the RPATH is
setup in a way that binaries can be used without issues for testing. The
only different platform here is Windows, which lack RPATH mechanism to
begin with.