Use `CONFIGURE_HANDLED_BY_BUILD` in superbuild
The CONFIGURE_HANDLED_BY_BUILD argument added in CMake 3.20 allows the
build system to skip calling the configure step of the child builds,
because they are also CMake builds, which are setup in a way that does
not make calling the configure step necessary. The build step is smart
enough to call the configure step on its own if its necessary, which
speeds up consequent calls to the superbuild's binary directory.

https://cmake.org/cmake/help/latest/module/ExternalProject.html#command:externalproject_add